pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smus-admin-agent"
version = "1.0.0"
description = "LangGraph-based AI agent powered by Anthropic Claude 4 Sonnet"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "SMUS Team", email = "smus@amazon.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langgraph>=0.0.40",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "mcp>=1.9.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
smus-admin-agent = "smus_admin_agent.cli:main"

[project.urls]
Homepage = "https://devcentral.amazon.com/ac/brazil/package-master/package/go/documentation?name=SMUS-data-agent"
Repository = "https://github.com/your-org/smus-data-agent"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100

[tool.isort]
known_first_party = ["smus_admin_agent"]

# required for compatibility with black:
profile = "black"

# To maintain consistency with other settings
line_length = 100

[tool.mypy]
# See https://mypy.readthedocs.io/en/latest/config_file.html for more mypy options.

# Enables the type-checker on the interior of functions without type annotations.
check_untyped_defs = true

# Displaying specific error codes makes it easier to silence specific errors
# See also https://mypy.readthedocs.io/en/latest/error_codes.html
show_error_codes = true

# Show source code snippets and location markers in error messages
pretty = true

# Suppresses errors about packages which do not implement type-hint sharing.
# See also https://mypy.readthedocs.io/en/latest/running_mypy.html#missing-imports.
ignore_missing_imports = true

[tool.pytest.ini_options]
xfail_strict = true
addopts = [
    "--verbose",
    "--cov=smus_admin_agent",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--durations=5",
    "--color=yes",
]
testpaths = ["test"]

